+++ FishingAssistant/ModEntry.cs
@@ -20,7 +20,7 @@ internal sealed class ModEntry : Mod
     public override void Entry(IModHelper helper)
     {
         this.configManager = new ConfigManager(helper, this.Monitor);
-        this.automationRuntime = new AutomationRuntime(this.Monitor);
+        this.automationRuntime = new AutomationRuntime(this.Monitor, () => this.configManager.Active);
         this.automationHud = new AutomationHudRenderer(key => helper.Translation.Get(key));
         this.starterFishingRod = new StarterFishingRodService(this.Monitor, key => helper.Translation.Get(key));
         ConfigValidationReport report = this.configManager.Load();